feat: Fullstory Analytics SDK Implementation (#2)
We have introduced the Fullstory Analytics Provider, which includes
three main methods:

Identify: This method identifies the user by passing a userID (uid). 
Additionally, it includes a displayName for use on the Fullstory 
dashboard.

Event: This method records custom app events.

Page: This method functions similarly to a screen event, tracking
page views.

There was an issue where the Fullstory cache was being cleared on
logout. To resolve this, we have added Fullstory to the exception list.

Fixes: LEARNER-9998